:sparkles: feat: Criei o service de excluir a movimentação

+++ server/src/modules/movement/movement.controller.ts
@@ -39,6 +39,9 @@ export class MovementController {
 
   @Delete(":id")
   remove(@Param("id") id: string) {
+    if(isNaN(+id)) {
+      throw new ConflictException("Id inválido");
+    }
     return this.movementService.remove(+id);
   }
 }
